fix(stats): route evergreen on-page counts through the live cache
Adds src/lib/library-stats.ts — a shared getLibraryStats() that reads the
daily-refreshed system_config.homepage_stats cache (books, languages,
illustrations, translations) plus a roundedCountLabel() formatter. One indexed
_id lookup; all callers are ISR/edge-cached, so no per-visit load.

Wires it into the surfaces that had hardcoded, drifted numbers:
- /collections hero — now uses the shared helper (was the inline fetch from
  the previous commit)
- /about/faq — "10,000+ books currently visible" -> live count
- /beta stats grid — split into a thin server page.tsx + BetaLandingClient.tsx
  so books/languages/illustrations render live in the initial HTML (no client
  fetch). These were badly stale: illustrations showed 74K+ vs ~191K actual,
  languages 100+ vs 162, books 10,000+ vs 16,328. Pages scanned/translated are
  page-level aggregates not in this cache — left as manual estimates.
